+++ internal/infra/web3/eth_gateway.go
@@ -1,8 +1,6 @@
 package web3
 
 import (
-	"context"
-	"log"
 	"math/big"
 	"os"
 	"strings"
@@ -10,8 +8,6 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/ethclient"
-	"github.com/orbit-alliance/orbit-backend/internal/domain/shared"
-	"github.com/orbit-alliance/orbit-backend/internal/domain/user"
 )
 
 type EthGateway struct {
@@ -47,24 +43,4 @@ func NewEthGateway(chainID int64) (*EthGateway, error) {
 		auth:            auth,
 		contractAddress: common.HexToAddress(contractAddr),
 	}, nil
-}
-
-func (g *EthGateway) PublishUserAction(ctx context.Context, payload *user.UserGoodAction) error {
-	tx, err := g.contract.PublishAction(
-		g.auth,
-		shared.ObjectIDToString(payload.ID),
-		common.HexToAddress(payload.UserWallet),
-		shared.ObjectIDToString(payload.UserID),
-		payload.Username,
-		shared.ObjectIDToString(payload.ActionID),
-		payload.ActionName,
-		big.NewInt(int64(payload.RewardAmount)),
-		big.NewInt(payload.PerformedAt.Unix()),
-	)
-	if err != nil {
-		return err
-	}
-
-	log.Printf("Transaction submitted: %s", tx.Hash().Hex())
-	return nil
 }
